+++ pkg/deployment/resources/pod_creator.go
@@ -547,7 +547,7 @@ func (r *Resources) createPodForMember(ctx context.Context, cachedStatus inspect
 		m.PodSpecVersion = template.PodSpecChecksum
 	}
 
-	member.GetPhaseExecutor().Execute(&m, api.Action{}, newPhase)
+	member.GetPhaseExecutor().Execute(r.context.GetAPIObject(), group, &m, api.Action{}, newPhase)
 
 	if top := status.Topology; top.Enabled() {
 		if m.Topology != nil && m.Topology.ID == top.ID {